Moved encryption to separate file and added encrypting into adding passwords.
